week 1-2/resume-analyser/requirements.txt
langchain==1.2.6
langchain-openai
langchain-community
pydantic
python-dotenv
pypdf
